- Update BackgroundTasks to override requireSession so that a session is actually provided when required  - Update IndexImageBackgroundTask to eagerly fetch license metdata to prevent Hibernate proxy error  - Update ClosableByteSource to delegate size methods  - Update ImageStoreService to pass the length parameter to the storage operation
