+++ src/main/groovy/au/org/ala/images/ScheduleLicenseReMatchAllBackgroundTask.groovy
@@ -3,7 +3,6 @@ package au.org.ala.images
 class ScheduleLicenseReMatchAllBackgroundTask extends BackgroundTask {
 
     private ImageService _imageService
-    boolean requiresSession = true
 
     ScheduleLicenseReMatchAllBackgroundTask(ImageService imageService) {
         _imageService = imageService
@@ -12,5 +11,10 @@ class ScheduleLicenseReMatchAllBackgroundTask extends BackgroundTask {
     @Override
     void execute() {
         _imageService.updateLicences()
+    }
+
+    @Override
+    boolean isRequiresSession() {
+        return true
     }
 }
